Color cycle for sprites
I didn't find a single color cycle with mask for unity so I modified the one which I found so it has a mask, so if you don't want to struggle with it here you go. I tested it in unity 2020.3.11f1. I using the universal render pipeline 10.5.0. Shader Graph 10.5.1. As of right now, I am not 100% sure if it needs shader graph, maybe it doesn't but I just write it down just in case.